requirements.txt
aiogram
aiohttp[speedups]
python-dotenv
fastapi[all]